// src/apis/AuthApi/types.ts
// Generated by https://quicktype.io

/**
 * LOGIN TYPE
 */

// Generated by https://quicktype.io

export interface LoginRequestType {
  email: string;
  password: string;
  deviceId: string;
}

export interface LoginResponseType {
  status: number;
  message: string;
  data: LoginResponseDataType;
}

export interface LoginResponseDataType {
  user: LoginResponseUserType;
  tokens: LoginResponseTokensType;
  deviceId: string;
}

export interface LoginResponseTokensType {
  access: LoginResponseTokenType;
  refresh: LoginResponseTokenType;
}

export interface LoginResponseTokenType {
  token: string;
  expires: string;
}

export interface LoginResponseUserType {
  id: number;
  username: string;
  email: string;
  password: string;
  contact: null;
  avatar: null;
  role: string;
  isEmailVerified: boolean;
  isContactVerified: boolean;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}

/**
 * REGISTER TYPE
 */

// Generated by https://quicktype.io

export interface RegisterRequestType {
  username: string;
  email: string;
  password: string;
}

// Generated by https://quicktype.io

export interface RegisterResponseType {
  status: number;
  message: string;
  data: RegisterResponseDataType;
}

export interface RegisterResponseDataType {
  user: RegisterResponseUserType;
}

export interface RegisterResponseUserType {
  contact: null;
  avatar: null;
  role: string;
  isEmailVerified: boolean;
  isContactVerified: boolean;
  isActive: boolean;
  id: number;
  username: string;
  email: string;
  password: string;
  updatedAt: string;
  createdAt: string;
}

/**
 * LOGOUT TYPE
 */

// Generated by https://quicktype.io

export interface LogOutRequestType {
  refreshToken: string;
  deviceId: string;
}

// Generated by https://quicktype.io

export interface LogOutResponseType {
  status: number;
  message: string;
  data: null;
}

/**
 * Refresh tokens type
 */

// Generated by https://quicktype.io

export interface RefreshTokensRequestType {
  refreshToken: string;
  deviceId: string;
}

// Generated by https://quicktype.io

export interface RefreshTokensResponseType {
  status: number;
  message: string;
  data: RefreshTokensResponseDataType;
}

export interface RefreshTokensResponseDataType {
  access: RefreshTokensResponseTokenType;
  refresh: RefreshTokensResponseTokenType;
}

export interface RefreshTokensResponseTokenType {
  token: string;
  expires: string;
}

/**
 * FORGOT PASSWORD TYPE
 */

// Generated by https://quicktype.io

export interface ForgotPasswordRequestType {
  email: string;
}

// Generated by https://quicktype.io

export interface ForgotPasswordResponseType {
  status: number;
  message: string;
  data: null;
}

/**
 * SEND VERIFICATION EMAIL TYPE
 */

// Generated by https://quicktype.io

export interface SendVerificationEmailRequestType {
  deviceId: string;
}
// Generated by https://quicktype.io

export interface SendVerificationEmailResponseType {
  status: number;
  message: string;
  data: null;
}

/**
 * Verify email type
 */

export interface VerifyEmailRequestType {
  deviceId: string;
}

export interface VerifyEmailResponseType {
  status: number;
  message: string;
  data: null;
}